+++ CHANGES.md
@@ -1,5 +1,9 @@
 ## Unreleased #unstable
 
+### Features
+
+ * opt-in for statistics collection (shared with TLC and TLA+ Toolbox), see #288
+
 ### Documentation
 
 - Compile the manuals into [a static